113548: created an api test fixture to share api test setup
